Implements MeasureType CoverPath
CoverPath is an unused MeasureType in the master branch. In this branch
it is implemented to be be a string output of the web address of the
AlbumArt. It is updated on finish of the downloading of the album art.
The intent is for users to be able to be able to use OnChange events on
measures of this type and not have to worry about the update rate of the
album art measure being too slow and the OnChange not firing or having
to deal with it double firing.