Fixed length and position seconds support
They were implmented as their own thing originally, now they are just
overloaded with the original values. Now in a number context it is just
the number and in the string context it is the string. I also left the
ms in for personal reasons, if you want to int cast it please make it
optional.